Small visual improvements for the notebook close button
Currently, because of the `notebook` and `sidebar` are not able to
communicate directly it is not possible to place the close button on the
`notebook` app itself.

While waiting for the above to be fixed, I propose a few minor style
changes:

* Changed the `LabeledButton` to `IconButton` and only show the `X`
  (cancel icon). This is a widely understood and accepted pattern for
  modals

* Increase the size of the cancel icon

* Re-enable the cursor when hovering on the button (it seems that the
  `cursor: pointer` rule was not longer functional)

* Made the margin bigger to avoid the browser scrolling bar to hidden by
  the button.

These changes are meant to be a temporary improvement.